# Improve fee estimation performance / fix accidentally quadratic function