--- a/lib/core/src/Cardano/Wallet/Primitive/Fee.hs
+++ b/lib/core/src/Cardano/Wallet/Primitive/Fee.hs
@@ -62,8 +62,6 @@ import Control.Monad.Trans.Except
     ( ExceptT (..), throwE )
 import Control.Monad.Trans.State
     ( StateT (..), evalStateT )
-import Crypto.Random.Types
-    ( MonadRandom )
 import Data.Word
     ( Word64 )
 import Fmt
@@ -169,11 +167,11 @@ newtype ErrAdjustForFee
 -- percentage of the fee (depending on how many change outputs the
 -- algorithm happened to choose).
 adjustForFee
-    :: (HasCallStack, MonadRandom m)
+    :: HasCallStack
     => FeeOptions
     -> UTxO
     -> CoinSelection
-    -> ExceptT ErrAdjustForFee m CoinSelection
+    -> ExceptT ErrAdjustForFee IO CoinSelection
 adjustForFee unsafeOpt utxo coinSel = do
     let opt = invariant "fee must be non-null" unsafeOpt (not . nullFee)
     cs <- senderPaysFee opt utxo coinSel
@@ -193,16 +191,14 @@ adjustForFee unsafeOpt utxo coinSel = do
 -- | The sender pays fee in this scenario, so fees are removed from the change
 -- outputs, and new inputs are selected if necessary.
 senderPaysFee
-    :: MonadRandom m
-    => FeeOptions
+    :: FeeOptions
     -> UTxO
     -> CoinSelection
-    -> ExceptT ErrAdjustForFee m CoinSelection
+    -> ExceptT ErrAdjustForFee IO CoinSelection
 senderPaysFee opt utxo sel = evalStateT (go sel) utxo where
     go
-        :: MonadRandom m
-        => CoinSelection
-        -> StateT UTxO (ExceptT ErrAdjustForFee m) CoinSelection
+        :: CoinSelection
+        -> StateT UTxO (ExceptT ErrAdjustForFee IO) CoinSelection
     go coinSel@(CoinSelection inps _ _ outs chgs _) = do
         -- Substract fee from change outputs, proportionally to their value.
         let (coinSel', remFee) = rebalanceSelection opt coinSel
@@ -239,9 +235,8 @@ senderPaysFee opt utxo sel = evalStateT (go sel) utxo where
 -- | A short / simple version of the 'random' fee policy to cover for fee in
 -- case where existing change were not enough.
 coverRemainingFee
-    :: MonadRandom m
-    => Fee
-    -> StateT UTxO (ExceptT ErrAdjustForFee m) [(TxIn, TxOut)]
+    :: Fee
+    -> StateT UTxO (ExceptT ErrAdjustForFee IO) [(TxIn, TxOut)]
 coverRemainingFee (Fee fee) = go [] where
     go acc
         | balance' acc >= fee =