--- a/lib/core/src/Cardano/Wallet/Primitive/Fee.hs
+++ b/lib/core/src/Cardano/Wallet/Primitive/Fee.hs
@@ -49,7 +49,6 @@ import Cardano.Wallet.Primitive.Types
     , TxIn
     , TxOut (..)
     , UTxO (..)
-    , balance'
     , invariant
     , isValidCoin
     , pickRandom
@@ -224,8 +223,8 @@ senderPaysFee opt utxo sel = evalStateT (go sel) utxo where
             -- re-run the algorithm with this new elements and using the initial
             -- change plus the extra change brought up by this entry and see if
             -- we can now correctly cover fee.
-            inps' <- coverRemainingFee remFee
-            let chgs' = splitChange (Coin $ balance' inps') chgs
+            (inps', surplus) <- coverRemainingFee remFee
+            let chgs' = splitChange surplus chgs
             go $ coinSel
                 { inputs  = inps <> inps'
                 , outputs = outs
@@ -236,18 +235,18 @@ senderPaysFee opt utxo sel = evalStateT (go sel) utxo where
 -- case where existing change were not enough.
 coverRemainingFee
     :: Fee
-    -> StateT UTxO (ExceptT ErrAdjustForFee IO) [(TxIn, TxOut)]
-coverRemainingFee (Fee fee) = go [] where
-    go acc
-        | balance' acc >= fee =
-            return acc
+    -> StateT UTxO (ExceptT ErrAdjustForFee IO) ([(TxIn, TxOut)], Coin)
+coverRemainingFee (Fee fee) = go [] 0 where
+    go additionalInputs surplus
+        | surplus >= fee =
+            return (additionalInputs, Coin surplus)
         | otherwise = do
             -- We ignore the size of the fee, and just pick randomly
             StateT (lift . pickRandom) >>= \case
-                Just entry ->
-                    go (entry : acc)
+                Just input@(_, out) ->
+                    go (input : additionalInputs) (getCoin (coin out) + surplus)
                 Nothing -> do
-                    lift $ throwE $ ErrCannotCoverFee (fee - balance' acc)
+                    lift $ throwE $ ErrCannotCoverFee (fee - surplus)
 
 -- | Reduce the given change outputs by the total fee, returning the remainig
 -- change outputs if any are left, or the remaining fee otherwise