# Issue Number

<!-- Put here a reference to the issue this PR relates to and which requirements it tackles -->

#2032 & rel

# Overview

<!-- Detail in a few bullet points the work accomplished in this PR -->

- ca9dd1ea2ded5b3539d7362f4fcb74fd2bcadf94
  :round_pushpin: **remove use of DerivingVia in Sqlite types: can't build with --profile with it**
  
- bc1bacc106faa4b17ad2a3cecf75eb29fb459bd6
  :round_pushpin: **avoid re-creating and folding on a map to compute balance of a list of resolved inputs**
    This is somehow causing massive slowness in the fee estimation. Changing this reduces the time in benchmarks from 18,000ms to 3,000ms!!

- f82a7fed3903f46409456c41ec46b292e2fa4bc6
  :round_pushpin: **use standard random generator for picking random UTxO**
    There's really no need for strong cryptographic randomness at this level, this is needlessly slower than it could be an called pretty often during fee calculation.

- 94d4812d459951fef3fa999008c35d9d781b4f0c
  :round_pushpin: **do not re-compute the full balance after picking new inputs**
    Instead, compute the balance incrementally. The 'coverRemainingFee' was computing the balance of the extra inputs selected to cover for fee. In worst-case scenario where many inputs needed to be selected, for example, 1000 extra inputs, it'll recompute the balance of an ever growing list a thousand times! Not only this is accidentally quadratic, but, it is also completely unbounded and will potentially select way too many inputs. In theory, this should not select more than the max allowed inputs. Next commit will fix this.

# Comments

<!-- Additional comments or screenshots to attach if any -->

Before:

```
Latency with 500 utxos scenario
    postTransactionFee  - 16756.3 ms
```

After:

```
    Latencies for 2 fixture wallets with 500 utxos scenario
        postTransactionFee  - 119.0 ms
```

<!-- 
Don't forget to:

 ✓ Self-review your changes to make sure nothing unexpected slipped through
 ✓ Assign yourself to the PR
 ✓ Assign one or several reviewer(s)
 ✓ Once created, link this PR to its corresponding ticket
 ✓ Assign the PR to a corresponding milestone
 ✓ Acknowledge any changes required to the Wiki
-->